Fix warnings misc

+++ tar/ccache/ccache_read.c
@@ -288,7 +288,7 @@ ccache_read(const char * path)
 		    (intmax_t)(fpos + (off_t)R.datalen));
 		goto err5;
 	}
-	C->datalen = R.datalen + (fpos % pagesize);
+	C->datalen = R.datalen + (size_t)(fpos % pagesize);
 	if ((C->data = mmap(NULL, C->datalen, PROT_READ,
 #ifdef MAP_NOCORE
 	    MAP_PRIVATE | MAP_NOCORE,
